+++ CHANGELOG.md
@@ -12,6 +12,7 @@ For a full diff see [`2.1.0...main`][2.1.0...main].
 
 - Dropped support for PHP 8.0 ([#393]), by [@localheinz]
 - Renamed `Parsed::fromFrontMatterAndContent()` to `Parsed::create()` ([#397]), by [@localheinz]
+- Renamed `Content` to `BodyMatter` and `Parsed::content()` to `Parsed::bodyMatter()` ([#398]), by [@localheinz]
 
 ## [`2.1.0`][2.1.0]